@@ -4,15 +4,11 @@
   "description": "frontend of easy market - ecommerce web app",
   "main": "index.js",
   "license": "MIT",
-  "engines": {
-    "node": ">=16.20.0"
-  },
   "scripts": {
     "dev": "next -p 7777",
-    "build-prod": "next build",
+    "build": "next build",
     "start": "next start -p 7777",
-    "lint": "next lint",
-    "test": "NODE_ENV=test jest --watch"
+    "lint": "next lint"
   },
   "dependencies": {
     "@apollo/client": "^3.7.9",
@@ -30,13 +26,13 @@
     "graphql-upload": "^16.0.2",
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
-    "next": "^12.3.2",
+    "next": "^13.4.3",
     "nprogress": "^0.2.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-transition-group": "^4.4.5",
-    "styled-components": "^5.3.9",
+    "styled-components": "^6.0.0-rc.1",
     "waait": "^1.0.5"
   },
   "devDependencies": {
