@@ -1,5 +1,8 @@
-module.exports = {
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   compiler: {
-    styledComponents: true
+      styledComponents: true
   }
 }
+
+module.exports = nextConfig
